golang: add callback method for http plugin config destruction (#38597)
This enables fine grained control over the lifecycle of golang filter
config in sync with C++.
Some use cases store states and resources in the config object that
needs to be cleaned when config is deleted or renewed.
The current design uses a Config interface, to minimise changes and
avoid breaking existing code.
I have an alternative design that adds a Destroy function in the
StreamFilterConfigParser interface instead of introducing an interface.
Let me know what you think, given the current go api is not considered
stable and breaking change should be acceptable.

Fixes #38557

Signed-off-by: François JACQUES <fjacques@murex.com>